feat(sorting): add ability to sort objects in ascending or descending order
Updated the `sortObject` function to include an additional parameter that allows objects to be sorted in either ascending or descending order. If the parameter is set to "asc", the objects will be sorted in ascending order; if it is set to "desc", they will be sorted in descending order.

Example usage:

```js
4 / 4

Here's the commit for the updated sortObject function:

vbnet
Copy code
feat(sorting): add ability to sort objects in ascending or descending order

Updated the `sortObject` function to include an additional parameter that allows objects to be sorted in either ascending or descending order. If the parameter is set to "asc", the objects will be sorted in ascending order; if it is set to "desc", they will be sorted in descending order.

Example usage:

const myObject = {
c: {
e: 'hello',
d: 'world'
},
b: 2,
a: {
d: 4,
c: 3,
b: 2,
a: 1
}
};

const sortedObject = sortObject(myObject, 'asc');

console.log(sortedObject);
// Output: {a: {a: 1, b: 2, c: 3, d: 4}, b: 2, c: {d: "world", e: "hello"}}

const sortedObjectDescending = sortObject(myObject, 'desc');

console.log(sortedObjectDescending);
// Output: {c: {e: "hello", d: "world"}, b: 2, a: {d: 4, c: 3, b: 2, a: 1}}
```